requirements-2023.txt
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy 
cycler 
fonttools==4.25.0
h5py==3.8.0
idna==3.4
importlib-resources 
kiwisolver==1.4.4
matplotlib==3.7.1
mkl-fft==1.3.6
mkl-random
mkl-service==2.4.0
munkres==1.1.4
numpy
opencv-python==4.7.0.72
packaging 
Pillow==9.4.0
ply==3.11
pyparsing
PyQt5-sip==12.11.0
python-dateutil 
requests==2.30.0
sip 
six 
toml
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tornado 
typing_extensions==4.5.0
urllib3==2.0.2
zipp
